backend/utils/jwt_handler.py
from datetime import datetime, timedelta, timezone
from typing import Optional, Dict, Any
from jose import jwt, JWTError

SECRET_KEY = "e4b213fcaf3b8dwqeqw7e4172391f92f94afc2b0e8f3f9e2324f48c57e2dc9eaa"
ALGORITHM = "HS256"

def create_access_token(data: dict, expires_hours: int = 2) -> str:
    to_encode = data.copy()
    expire = datetime.now(timezone.utc) + timedelta(hours=expires_hours)
    to_encode.update({"exp": expire})
    return jwt.encode(to_encode, SECRET_KEY, algorithm=ALGORITHM)

def decode_token(token: str) -> Optional[Dict[str, Any]]:
    try:
        payload = jwt.decode(token, SECRET_KEY, algorithms=[ALGORITHM])
        return payload
    except JWTError:
        return None
