backend/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
cffi==2.0.0
click==8.1.8
cryptography==46.0.2
ecdsa==0.19.1
exceptiongroup==1.3.0
fastapi==0.118.0
h11==0.16.0
idna==3.10
pyasn1==0.6.1
pycparser==2.23
pydantic==2.11.10
pydantic_core==2.33.2
email-validator==2.2.0
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.37.0
